Create school management system
this is a group assignment. 
CP/H Dhammaloka vidyalaya is a small scale type 02 school situated in Badupola,Ginigathhena near by the Colombo-Hatton main road in central province. 

The school and the staff has been awarded and appriciated several times because of their productivity and their quality service to students and community even though the school is categorized under minimum resourced schools. 

Under this circumstances the principal of the school Mrs. P.H.M. Mangalika decided to establish a student information system in-order to improve the efficiency of the staff and to provide quick and quality service to students and relevant parties. 

As the staff and the principal mentioned they met with following difficulties due to absence of such a system in the school. 

* It cosume a lot of time to prepare the documents that related to students since all details are recorded in different documets and have to prepare those documents by type setting from computer again and again. 

*It's difficult to find some essential details such as contact details in emergency cases.
Accordingly we were requested to conduct an analysis on the above mentioned scinario and to propose the most suitable student  information system that meets with their requirements.